Add files via upload
"This is a picture of a program that I made for our automated liquid handling system at Quest. Essentially, this program tells a robot to pick up pipette tips from one location, aspirate a solvent from another location, and dispense that solvent into a final location. Since the robot is able to do this 96 at a time (the "probe" contains 96 compact pipetting channels, corresponding to the 96-well plate used in analytical chemistry laboratories), it saves time while doing the work with accuracy and precision. The program was made through a GUI, but can also be done via traditional coding. The program allows the user to select a desired number of aspiration-dispensing cycles (1-4), corresponding to the number of plates that the user would like to have solvent added to. As you can see, this is done via if-else statements, where var_NumPltsTotal dictates which if-else path is taken. If var_NumPltsTotal is assigned a value of 1 by the user, then the machine will only aspirate and dispense solvent one time; otherwise, the program will move along the elif arguments until var_NumPltsTotal finds a matching value. This structure is very similar to the if, else statements we learned in the Python Zybooks assignments, and I just thought it was a cool thing to share with the class."